reintroduce-access-scope-validation
