+++ central/role/validate.go
@@ -9,9 +9,8 @@ import (
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/auth/permissions"
 	"github.com/stackrox/rox/pkg/env"
-	"github.com/stackrox/rox/pkg/sac/effectiveaccessscope"
+	"github.com/stackrox/rox/pkg/role"
 	"github.com/stackrox/rox/pkg/uuid"
-	"k8s.io/apimachinery/pkg/labels"
 )
 
 const (
@@ -152,50 +151,10 @@ func ValidateSimpleAccessScope(scope *storage.SimpleAccessScope) error {
 		multiErr = multierror.Append(multiErr, errors.New("name field must be set"))
 	}
 
-	err := ValidateSimpleAccessScopeRules(scope.GetRules())
+	err := role.ValidateSimpleAccessScopeRules(scope.GetRules())
 	if err != nil {
 		multiErr = multierror.Append(multiErr, err)
 	}
 
-	return multiErr
-}
-
-// ValidateSimpleAccessScopeRules checks whether the supplied protobuf message
-// represents valid simple access scope rules.
-func ValidateSimpleAccessScopeRules(scopeRules *storage.SimpleAccessScope_Rules) error {
-	var multiErr error
-
-	for _, ns := range scopeRules.GetIncludedNamespaces() {
-		if ns.GetClusterName() == "" || ns.GetNamespaceName() == "" {
-			multiErr = multierror.Append(multiErr, errors.Errorf(
-				"both cluster_name and namespace_name fields must be set in namespace rule <%s, %s>",
-				ns.GetClusterName(), ns.GetNamespaceName()))
-		}
-	}
-	for _, labelSelector := range scopeRules.GetClusterLabelSelectors() {
-		err := validateSelectorRequirement(labelSelector)
-		if err != nil {
-			multiErr = multierror.Append(multiErr, err)
-		}
-	}
-	for _, labelSelector := range scopeRules.GetNamespaceLabelSelectors() {
-		err := validateSelectorRequirement(labelSelector)
-		if err != nil {
-			multiErr = multierror.Append(multiErr, err)
-		}
-	}
-
-	return multiErr
-}
-
-func validateSelectorRequirement(labelSelector *storage.SetBasedLabelSelector) error {
-	var multiErr error
-	for _, requirement := range labelSelector.GetRequirements() {
-		op := effectiveaccessscope.ConvertLabelSelectorOperatorToSelectionOperator(requirement.GetOp())
-		_, err := labels.NewRequirement(requirement.GetKey(), op, requirement.Values)
-		if err != nil {
-			multiErr = multierror.Append(multiErr, err)
-		}
-	}
 	return multiErr
 }
